# specs/api/openapi.yaml
openapi: 3.1.0
info:
  title: YOLO-Toys API
  description: |
    Multi-model real-time vision recognition platform supporting YOLO, HuggingFace Transformers,
    and multimodal models via WebSocket streaming.

    ## Features

    - 🎯 Object Detection (YOLOv8, DETR)
    - 🖼️ Instance Segmentation (YOLOv8-seg)
    - 🏃 Pose Estimation (YOLOv8-pose)
    - 🔍 Open-Vocabulary Detection (OWL-ViT, Grounding DINO)
    - 📝 Image Captioning (BLIP)
    - 💬 Visual Question Answering (BLIP VQA)

    ## WebSocket Streaming

    For real-time inference, use WebSocket endpoint:

    ```
    ws://localhost:8000/ws?model=yolov8n.pt&conf=0.25&iou=0.45
    ```
  version: 3.1.0
  contact:
    name: YOLO-Toys Team
    url: https://github.com/LessUp/yolo-toys
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8000
    description: Local development server

tags:
  - name: Health
    description: System health and status
  - name: Models
    description: Model discovery and information
  - name: Inference
    description: Image inference endpoints
  - name: Multimodal
    description: Multimodal AI endpoints

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      description: Check server status and system information
      operationId: getHealth
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: ok
                version: 3.1.0
                device: cuda:0
                default_model: yolov8n.pt
                defaults:
                  conf: 0.25
                  iou: 0.45
                  max_det: 300

  /models:
    get:
      tags: [Models]
      summary: List available models
      description: Get all available models grouped by category
      operationId: listModels
      responses:
        '200':
          description: List of models by category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelsResponse'

  /models/{model_id}:
    get:
      tags: [Models]
      summary: Get model details
      description: Get detailed information about a specific model
      operationId: getModel
      parameters:
        - name: model_id
          in: path
          required: true
          description: Model identifier
          schema:
            type: string
          example: yolov8n.pt
      responses:
        '200':
          description: Model details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelInfo'
        '404':
          description: Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /labels:
    get:
      tags: [Models]
      summary: Get model labels
      description: Get class labels for a model
      operationId: getLabels
      parameters:
        - name: model
          in: query
          required: false
          description: Model identifier (defaults to configured model)
          schema:
            type: string
            default: yolov8n.pt
      responses:
        '200':
          description: Model labels
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelsResponse'

  /infer:
    post:
      tags: [Inference]
      summary: Single image inference
      description: |
        Perform inference on a single image. Supports:

        - Object detection (YOLO, DETR)
        - Instance segmentation (YOLO-seg)
        - Pose estimation (YOLO-pose)
        - Open-vocabulary detection (OWL-ViT, Grounding DINO)
      operationId: infer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/InferRequest'
      responses:
        '200':
          description: Inference result
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/DetectionResult'
                  - $ref: '#/components/schemas/SegmentationResult'
                  - $ref: '#/components/schemas/PoseResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Inference error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /caption:
    post:
      tags: [Multimodal]
      summary: Image captioning
      description: Generate automatic description for an image
      operationId: caption
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CaptionRequest'
      responses:
        '200':
          description: Generated caption
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptionResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /vqa:
    post:
      tags: [Multimodal]
      summary: Visual question answering
      description: Answer questions about an image
      operationId: vqa
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/VQARequest'
      responses:
        '200':
          description: Answer to the question
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VQAResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    HealthResponse:
      type: object
      required: [status, version, device, default_model, defaults]
      properties:
        status:
          type: string
          enum: [ok, error]
        version:
          type: string
        device:
          type: string
        default_model:
          type: string
        defaults:
          type: object
          properties:
            conf:
              type: number
              format: float
            iou:
              type: number
              format: float
            max_det:
              type: integer

    ModelsResponse:
      type: object
      properties:
        categories:
          type: object
          additionalProperties:
            type: object
            properties:
              name:
                type: string
              models:
                type: array
                items:
                  $ref: '#/components/schemas/ModelSummary'

    ModelSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        speed:
          type: string
        accuracy:
          type: string

    ModelInfo:
      type: object
      properties:
        id:
          type: string
        category:
          type: string
        name:
          type: string
        description:
          type: string
        task:
          type: string
          enum: [detect, segment, pose, caption, vqa]
        parameters:
          type: object
          additionalProperties:
            type: object
            properties:
              default:
                type: [number, integer, string, boolean]
              min:
                type: number
              max:
                type: number

    LabelsResponse:
      type: object
      properties:
        model:
          type: string
        labels:
          type: array
          items:
            type: string

    InferRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: Image file (JPEG, PNG, WEBP)
        model:
          type: string
          default: yolov8n.pt
          description: Model identifier
        conf:
          type: number
          format: float
          default: 0.25
          minimum: 0.0
          maximum: 1.0
          description: Confidence threshold
        iou:
          type: number
          format: float
          default: 0.45
          minimum: 0.0
          maximum: 1.0
          description: IoU threshold for NMS
        max_det:
          type: integer
          default: 300
          minimum: 1
          maximum: 1000
          description: Maximum detections
        device:
          type: string
          default: auto
          description: Inference device (cpu/cuda/mps)
        imgsz:
          type: integer
          default: 640
          minimum: 32
          maximum: 4096
          description: Inference image size
        half:
          type: boolean
          default: false
          description: Enable FP16 half-precision
        text_queries:
          type: string
          description: Text queries for open-vocabulary detection (comma-separated)
        question:
          type: string
          maxLength: 500
          description: Question for VQA models

    Detection:
      type: object
      properties:
        bbox:
          type: array
          items:
            type: number
            format: float
          minItems: 4
          maxItems: 4
          description: Bounding box [x1, y1, x2, y2] in pixel coordinates
        score:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Detection confidence score
        label:
          type: string
          description: Class label
        track_id:
          type: integer
          nullable: true
          description: Tracking ID (if tracking enabled)
        polygons:
          type: array
          items:
            type: array
            items:
              type: array
              items:
                type: number
              minItems: 2
              maxItems: 2
          description: Segmentation polygons (for segment task)
        keypoints:
          type: array
          items:
            type: object
            properties:
              x:
                type: number
              y:
                type: number
              confidence:
                type: number
                format: float
          description: Keypoints (for pose task)

    DetectionResult:
      type: object
      properties:
        width:
          type: integer
          description: Image width in pixels
        height:
          type: integer
          description: Image height in pixels
        task:
          type: string
          enum: [detect]
        detections:
          type: array
          items:
            $ref: '#/components/schemas/Detection'
        inference_time:
          type: number
          format: float
          description: Inference time in milliseconds
        model:
          type: string
          description: Model identifier used

    SegmentationResult:
      type: object
      properties:
        width:
          type: integer
        height:
          type: integer
        task:
          type: string
          enum: [segment]
        detections:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/Detection'
              - type: object
                required: [polygons]
        inference_time:
          type: number
          format: float
        model:
          type: string

    PoseResult:
      type: object
      properties:
        width:
          type: integer
        height:
          type: integer
        task:
          type: string
          enum: [pose]
        detections:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/Detection'
              - type: object
                required: [keypoints]
        inference_time:
          type: number
          format: float
        model:
          type: string

    CaptionRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: Image file
        model:
          type: string
          default: Salesforce/blip-image-captioning-base
          description: Caption model identifier

    CaptionResult:
      type: object
      properties:
        caption:
          type: string
          description: Generated image caption
        inference_time:
          type: number
          format: float
          description: Inference time in milliseconds
        model:
          type: string
          description: Model identifier used

    VQARequest:
      type: object
      required: [file, question]
      properties:
        file:
          type: string
          format: binary
          description: Image file
        question:
          type: string
          maxLength: 500
          description: Question about the image
        model:
          type: string
          default: Salesforce/blip-vqa-base
          description: VQA model identifier

    VQAResult:
      type: object
      properties:
        answer:
          type: string
          description: Answer to the question
        question:
          type: string
          description: The asked question
        inference_time:
          type: number
          format: float
          description: Inference time in milliseconds
        model:
          type: string
          description: Model identifier used

    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message

tags: []